modify: Move add menu endpoint
